changed fragment structure
