+++ app/src/main/res/layout/fragment_view_profile.xml
@@ -8,10 +8,10 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context=".ui.home.HomeFragment"
-    android:id="@+id/layout_assistant_home">
+    android:id="@+id/layout_view_profile">
 
     <FrameLayout
-        android:id="@+id/assistant_home_frame"
+        android:id="@+id/home_frame"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
