+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -37,8 +37,8 @@
 
     <fragment
         android:id="@+id/nav_assistant_home"
-        android:name="com.courseproject.tindar.ui.home.AssistantHomeFragment"
+        android:name="com.courseproject.tindar.ui.home.SecondViewProfileFragment"
         android:label="View Profile"
-        tools:layout="@layout/fragment_assistant_home" />
+        tools:layout="@layout/fragment_second_view_profile" />
 
 </navigation>
